dev-requirements.txt
vcrpy==4.1.1
pytest==6.2.2
pytest-sugar==0.9.4
msrestazure==0.6.4
azure-mgmt-storage==17.0.0
